Modified plan, Added directions of research
I deleted all references to "scraping/downloading Wikipedia" because there's an API; it is possible that downloading a bunch will be faster because we're running things locally, but it feels to me like scraping it ourselves will probably not be better than downloading it from some dump.  Feel free to change this back.